requirements.txt
numpy==1.18.1
matplotlib==3.1.3
pandas==1.0.5
seaborn==0.10.0
Flask==1.1.1
scikit-learn==0.23.1
gunicorn==20.0.4
joblib==0.14.1
Pillow